fix(arch): relocate SearchBackend to a new L1 ai_provider module (R7)
Pre-push architecture gate caught it: cover_letter/research/mod.rs (L2)
importing commands::ai_provider::search::SearchBackend (L3) is an
upward layer dependency, introduced by the round-2 search-backend key
term. pipeline/mod.rs's Completer::search_backend() had the identical
upward reference in its return-type annotation (not flagged by the
gate, since pipeline->commands is separately allowlisted debt (W-1),
but the same shape).

Fixed structurally, not with an allowlist entry: split the pure
SearchBackend enum + as_str() out of commands::ai_provider::search
into a new top-level src/ai_provider/ module, classified L1 in both
tests/architecture.rs and docs/architecture-rules.md (a deliberately
minimal seed of the roadmap's "relocate ai_provider out of commands/"
item -- only the value type moved; search_backend_for's credential/
AppHandle-reading resolution logic stays L3, in commands/ai_provider,
which now `pub use`s the relocated type so its existing consumers
(resolve_search_backend, searcher_for, search_backend_for, and the
search/test.rs suite via `use super::*`) are unaffected). Both L2
call sites (cover_letter::research, pipeline::Completer::search_backend)
now import crate::ai_provider::SearchBackend directly -- L2 depending
on L1, no exception needed.

Verified: cargo test --lib --all-features (4137 passed), cargo test
--test architecture (18/18, including r7_no_upward_layer_imports and
r8_no_oversized_modules -- the latter's earlier failure was a
concurrent, in-progress export/typst_engine/letter.rs split by
another agent in this shared worktree, unrelated to this change and
resolved on its own once that split completed), clippy --all-targets
--all-features --locked -D warnings (clean), fmt (clean).
